chore: Undo bumping WindowedTableSource execution step version (#8111)
