+++ ksqldb-engine/src/main/java/io/confluent/ksql/util/PlanSummary.java
@@ -48,7 +48,6 @@ import io.confluent.ksql.execution.plan.TableSuppress;
 import io.confluent.ksql.execution.plan.TableTableJoin;
 import io.confluent.ksql.execution.plan.WindowedStreamSource;
 import io.confluent.ksql.execution.plan.WindowedTableSource;
-import io.confluent.ksql.execution.plan.WindowedTableSourceV1;
 import io.confluent.ksql.execution.streams.StepSchemaResolver;
 import io.confluent.ksql.metastore.MetaStore;
 import io.confluent.ksql.query.QueryId;
@@ -100,7 +99,6 @@ public class PlanSummary {
           .put(TableSourceV1.class, "SOURCE")
           .put(TableSource.class, "SOURCE")
           .put(TableSuppress.class, "SUPPRESS")
-          .put(WindowedTableSourceV1.class, "SOURCE")
           .put(WindowedTableSource.class, "SOURCE")
           .build();
