+++ ksqldb-execution/src/main/java/io/confluent/ksql/execution/plan/ExecutionStep.java
@@ -45,8 +45,7 @@ import javax.annotation.Nonnull;
     @Type(value = StreamWindowedAggregate.class, name = "streamWindowedAggregateV1"),
     @Type(value = TableSourceV1.class, name = "tableSourceV1"),
     @Type(value = TableSource.class, name = "tableSourceV2"),
-    @Type(value = WindowedTableSourceV1.class, name = "windowedTableSourceV1"),
-    @Type(value = WindowedTableSource.class, name = "windowedTableSourceV2"),
+    @Type(value = WindowedTableSource.class, name = "windowedTableSourceV1"),
     @Type(value = TableAggregate.class, name = "tableAggregateV1"),
     @Type(value = TableFilter.class, name = "tableFilterV1"),
     @Type(value = TableGroupByV1.class, name = "tableGroupByV1"),
